requirements.txt
async-timeout==4.0.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
loguru==0.7.2
MarkupSafe==2.1.5
packaging==24.1
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
telebot==0.0.5
urllib3==2.2.2
Werkzeug==3.0.3
win32-setctime==1.1.0
